requirements.txt
datasets
evaluate
sentencepiece
accelerate
bitsandbytes
ipywidgets
protobuf==4.25.3
google-generativeai==0.5.4
langchain==0.2.1
langchain-community==0.2.1
langchain-openai
langchain-google-genai==1.0.5
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
Werkzeug==2.0.2
python-dotenv==1.0.1
llama-index==0.10.30
llama-index-vector-stores-milvus
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-embeddings-langchain==0.1.2
llama-index-llms-huggingface
llama-index-llms-gemini==0.1.9
pdfminer.six==20231228
streamlit==1.34.0
docx2txt==0.8
openpyxl==3.1.2
aspose-cells==24.5.0
unstructured[all-docs]==0.14.2
InstructorEmbedding==1.0.1
langchain-pinecone==0.1.1
llama-index-readers-wikipedia==0.1.4
llama-hub==0.0.79.post1
llama-index-embeddings-instructor==0.1.3
wikipedia==1.4.0
comtypes==1.4.2
aspose-words==24.5.0
llama-index-llms-cohere==0.2.0
llama-index-postprocessor-cohere-rerank==0.1.6
llama-index-vector-stores-pinecone==0.1.7
arize-phoenix[evals]
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
ipython==7.29.0
Werkzeug==2.0.2